restructure app to be more modular
