+++ xbox/rest/routes/__init__.py
@@ -1,0 +1,8 @@
+from flask import Blueprint
+routes = Blueprint('routes', __name__)
+
+from .index import *
+from .auth import *
+from .device import *
+from .web import *
+from .versions import *
