+++ xbox/rest/routes/device.py
@@ -1,0 +1,266 @@
+from flask import current_app as app
+from flask import request
+from xbox.sg import enum
+from ..decorators import console_exists, console_connected
+from ..consolewrap import ConsoleWrap
+from . import routes
+
+@routes.route('/device')
+def device_overview():
+    discovered = ConsoleWrap.discover().copy()
+
+    liveids = [d.liveid for d in discovered]
+    for i, c in enumerate(app.console_cache.values()):
+        if c.liveid in liveids:
+            # Refresh existing entries
+            index = liveids.index(c.liveid)
+
+            if c.device_status != discovered[index].device_status:
+                app.console_cache[c.liveid] = ConsoleWrap(discovered[index])
+            del discovered[index]
+            del liveids[index]
+        elif c.liveid not in liveids:
+            # Set unresponsive consoles to Unavailable
+            app.console_cache[c.liveid].console.device_status = enum.DeviceStatus.Unavailable
+
+    # Extend by new entries
+    for d in discovered:
+        app.console_cache.update({d.liveid: ConsoleWrap(d)})
+
+    data = {console.liveid: console.status for console in app.console_cache.values()}
+    return app.success(devices=data)
+
+
+@routes.route('/device/<liveid>/poweron')
+def poweron(liveid):
+    ConsoleWrap.power_on(liveid)
+    return app.success()
+
+
+"""
+Require enumerated console
+"""
+
+
+@routes.route('/device/<liveid>')
+@console_exists
+def device_info(console):
+    return app.success(device=console.status)
+
+
+@routes.route('/device/<liveid>/connect', methods=['GET', 'POST'])
+@console_exists
+def force_connect(console):
+    connect_anonymous = False
+
+    if request.method == 'POST':
+        connect_anonymous = True if request.form.get('connect_anonymous') else False
+
+    try:
+        state = console.connect(connect_anonymous=connect_anonymous)
+    except Exception as e:
+        return app.error(str(e))
+
+    if state != enum.ConnectionState.Connected:
+        return app.error('Connection failed', connection_state=state.name)
+
+    return app.success(connection_state=state.name)
+
+
+"""
+Require connected console
+"""
+
+
+@routes.route('/device/<liveid>/disconnect')
+@console_connected
+def disconnect(console):
+    console.disconnect()
+    return app.success()
+
+
+@routes.route('/device/<liveid>/poweroff')
+@console_connected
+def poweroff(console):
+    if not console.power_off():
+        return app.error("Failed to power off")
+    else:
+        return app.success()
+
+
+@routes.route('/device/<liveid>/console_status')
+@console_connected
+def console_status(console):
+    return app.success(console_status=console.console_status)
+
+
+@routes.route('/device/<liveid>/launch/<app_id>')
+@console_connected
+def launch_title(console, app_id):
+    console.launch_title(app_id)
+    return app.success(launched=app_id)
+
+
+@routes.route('/device/<liveid>/media_status')
+@console_connected
+def media_status(console):
+    return app.success(media_status=console.media_status)
+
+
+@routes.route('/device/<liveid>/ir')
+@console_connected
+def infrared(console):
+    stump_config = console.stump_config
+
+    devices = {}
+    for device_config in stump_config.params:
+        button_links = {}
+        for button in device_config.buttons:
+            button_links[button] = {
+                'url': '/device/{0}/ir/{1}/{2}'.format(console.liveid, device_config.device_id, button),
+                'value': device_config.buttons[button]
+            }
+
+        devices[device_config.device_type] = {
+            'device_type': device_config.device_type,
+            'device_brand': device_config.device_brand,
+            'device_model': device_config.device_model,
+            'device_name': device_config.device_name,
+            'device_id': device_config.device_id,
+            'buttons': button_links
+        }
+
+    return app.success(**devices)
+
+
+@routes.route('/device/<liveid>/ir/<device_id>')
+@console_connected
+def infrared_available_keys(console, device_id):
+    stump_config = console.stump_config
+    for device_config in stump_config.params:
+        if device_config.device_id != device_id:
+            continue
+
+        button_links = {}
+        for button in device_config.buttons:
+            button_links[button] = {
+                'url': '/device/{0}/ir/{1}/{2}'.format(console.liveid, device_config.device_id, button),
+                'value': device_config.buttons[button]
+            }
+
+        return app.success(
+            device_type=device_config.device_type,
+            device_brand=device_config.device_brand,
+            device_model=device_config.device_model,
+            device_name=device_config.device_name,
+            device_id=device_config.device_id,
+            buttons=button_links
+        )
+
+    return app.error('Device Id \'{0}\' not found'.format(device_id))
+
+
+@routes.route('/device/<liveid>/ir/<device_id>/<button>')
+@console_connected
+def infrared_send(console, device_id, button):
+    if not console.send_stump_key(device_id, button):
+        return app.error('Failed to send button')
+
+    return app.success(sent_key=button, device_id=device_id)
+
+
+@routes.route('/device/<liveid>/media')
+@console_connected
+def media_overview(console):
+    commands = [cmd.name for cmd in enum.MediaControlCommand]
+    return app.success(commands=commands)
+
+
+@routes.route('/device/<liveid>/media/<command>')
+@console_connected
+def media_command(console, command):
+    try:
+        cmd = enum.MediaControlCommand[command]
+    except Exception as e:
+        return app.error('Invalid command passed, msg: {0}'.format(e))
+
+    console.send_media_command(cmd)
+    return app.success()
+
+
+@routes.route('/device/<liveid>/media/Seek/<seek_position>')
+@console_connected
+def media_command_seek(console, seek_pos):
+    console.send_media_command(enum.MediaControlCommand.Seek, int(seek_pos))
+    return app.success()
+
+
+@routes.route('/device/<liveid>/input')
+@console_connected
+def input_overview(console):
+    buttons = [btn.name for btn in enum.GamePadButton]
+    return app.success(buttons=buttons)
+
+
+@routes.route('/device/<liveid>/input/<button>')
+@console_connected
+def input_send_button(console, button):
+    try:
+        btn = enum.GamePadButton[button]
+    except Exception as e:
+        return app.error('Invalid button passed, msg: {0}'.format(e))
+
+    console.send_gamepad_button(btn)
+    return app.success()
+
+
+@routes.route('/device/<liveid>/stump/headend')
+@console_connected
+def stump_headend_info(console):
+    return app.success(headend_info=console.headend_info.params.dump())
+
+
+@routes.route('/device/<liveid>/stump/livetv')
+@console_connected
+def stump_livetv_info(console):
+    return app.success(livetv_info=console.livetv_info.params.dump())
+
+
+@routes.route('/device/<liveid>/stump/tuner_lineups')
+@console_connected
+def stump_tuner_lineups(console):
+    return app.success(tuner_lineups=console.tuner_lineups.params.dump())
+
+
+@routes.route('/device/<liveid>/text')
+@console_connected
+def text_overview(console):
+    return app.success(text_session_active=console.text_active)
+
+
+@routes.route('/device/<liveid>/text/<text>')
+@console_connected
+def text_send(console, text):
+    console.send_text(text)
+    return app.success()
+
+"""
+@routes.route('/device/<liveid>/nano')
+@console_connected
+def nano_overview(console):
+    return app.success(nano_status=console.nano_status)
+
+
+@routes.route('/device/<liveid>/nano/start')
+@console_connected
+def nano_start(console):
+    console.nano_start()
+    return app.success()
+
+
+@routes.route('/device/<liveid>/nano/stop')
+@console_connected
+def nano_stop(console):
+    console.nano_stop()
+    return app.success()
+"""
